+++ importlib_metadata/api.py
@@ -68,12 +68,21 @@ class Distribution:
         metadata.  See PEP 566 for details.
         """
         text = self.read_text('METADATA') or self.read_text('PKG-INFO')
-        return email.message_from_string(text)
+        return _email_message_from_string(text)
 
     @property
     def version(self):
         """Return the 'Version' metadata for the distribution package."""
         return self.metadata['Version']
+
+
+def _email_message_from_string(text):
+    # Work around https://bugs.python.org/issue25545 where
+    # email.message_from_string cannot handle Unicode on Python 2.
+    if sys.version_info < (3,):                     # nocoverpy3
+        io_buffer = io.StringIO(text)
+        return email.message_from_file(io_buffer)
+    return email.message_from_string(text)          # nocoverpy2
 
 
 def distribution(package):